+++ app/Http/Controllers/CashOutController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\CashOut;
+use App\Models\Fund;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CashOutController extends Controller
 {
@@ -22,7 +24,8 @@ class CashOutController extends Controller
      */
     public function create()
     {
-        return view("dashboard.finance.cash-out.create");
+        $funds = Fund::where('company_id', Auth::user()->company_id)->get();
+        return view("dashboard.finance.cash-out.create", ["funds" => $funds]);
     }
 
     /**
@@ -34,18 +37,32 @@ class CashOutController extends Controller
             'fund' => 'required',
             'remark' => 'required',
             'datetime' => 'required',
+            'type' => 'required',
         ]);
 
         $store = CashOut::create([
-            'company_id' => 1,
+            'company_id' => Auth::user()->company_id,
             'fund' => $validate['fund'],
             'remark' => $validate['remark'],
             'datetime' => $validate['datetime'],
+            'type' => $validate['type'],
         ]);
 
         $query_data = ['periode' => Carbon::parse($validate['datetime'])->format('Y-m-d')];
 
         if ($store) {
+            $fund = Fund::where(
+                "company_id",
+                Auth::user()->company_id
+            )->where(
+                "type",
+                $request["type"]
+            )->first();
+
+            $fund->update([
+                "fund" => $fund->fund - $validate["fund"]
+            ]);
+
             return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('success', "Successfully to create cash out");
         } else {
             return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to create cash out");
@@ -58,7 +75,16 @@ class CashOutController extends Controller
     public function edit(string $id)
     {
         $data = CashOut::findOrFail($id);
-        return view('dashboard.finance.cash-out.edit', ["data" => $data]);
+        $funds = Fund::where('company_id', Auth::user()->company_id)->get();
+
+        if ($data->company_id == Auth::user()->company_id) {
+            return view('dashboard.finance.cash-out.edit', [
+                "data" => $data,
+                "funds" => $funds,
+            ]);
+        } else {
+            return view('dashboard.finance.cash-out')->with('failed', 'Oops! Looks like you followed a bad link. If you think this is a problem with us, please tell us.');
+        }
     }
 
     /**
@@ -70,23 +96,64 @@ class CashOutController extends Controller
             'fund' => 'required',
             'remark' => 'required',
             'datetime' => 'required',
+            'type' => 'required',
         ]);
 
         $data = CashOut::findOrFail($id);
+        $type = $data->type;
+        $amount = $data->fund;
 
-        $update = $data->update([
-            'company_id' => 1,
-            'fund' => $validate['fund'],
-            'remark' => $validate['remark'],
-            'datetime' => $validate['datetime'],
-        ]);
+        if ($data->company_id == Auth::user()->company_id) {
 
-        $query_data = ['periode' => Carbon::parse($validate['datetime'])->format('Y-m-d')];
+            $update = $data->update([
+                'company_id' => Auth::user()->company_id,
+                'fund' => $validate['fund'],
+                'remark' => $validate['remark'],
+                'datetime' => $validate['datetime'],
+                'type' => $validate['type'],
 
-        if ($update) {
-            return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('success', "Successfully to update cash out");
+            ]);
+
+            $query_data = ['periode' => Carbon::parse($validate['datetime'])->format('Y-m-d')];
+
+            if ($update) {
+                if ($type == $validate["type"]) {
+                    $fund = Fund::where(
+                        "company_id",
+                        Auth::user()->company_id
+                    )->where('type', $type)->first();
+
+                    if ($fund != $validate["fund"]) {
+                        $fund->update([
+                            "fund" => $fund->fund + $amount - $validate["fund"]
+                        ]);
+                    }
+                } else {
+                    $fundOld = Fund::where(
+                        "company_id",
+                        Auth::user()->company_id
+                    )->where('type', $type)->first();
+
+                    $fund = Fund::where(
+                        "company_id",
+                        Auth::user()->company_id
+                    )->where('type', $validate['type'])->first();
+
+                    $fundOld->where('type', $type)->update([
+                        "fund" => $fundOld->fund + $amount
+                    ]);
+
+                    $fund->where('type', $validate['type'])->update([
+                        "fund" => $fund->fund - $validate["fund"]
+                    ]);
+                }
+
+                return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('success', "Successfully to update cash out");
+            } else {
+                return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to update cash out");
+            }
         } else {
-            return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to update cash out");
+            return abort(404);
         }
     }
 
@@ -100,11 +167,16 @@ class CashOutController extends Controller
             $query_data = ['periode' => $request->periode];
         }
 
-        $delete =  CashOut::destroy($id);
-        if ($delete) {
-            return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('success', "Successfully to delete cash out");
+        $data = CashOut::findOrFail($id);
+        if ($data->company_id == Auth::user()->company_id) {
+            $delete =  CashOut::destroy($id);
+            if ($delete) {
+                return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('success', "Successfully to delete cash out");
+            } else {
+                return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to delete cash out");
+            }
         } else {
-            return redirect()->route('dashboard.finance.cash-flow-daily', $query_data)->with('failed', "Failed to delete cash out");
+            return abort(404);
         }
     }
 }
